requirements.txt
astroid==2.2.5
atomicwrites==1.4.0
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
certifi==2020.12.5
Click==7.0
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
ecdsa==0.13.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.0
future==0.17.1
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
iniconfig @ file:///tmp/build/80754af9/iniconfig_1602780191262/work
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools @ file:///tmp/build/80754af9/more-itertools_1605111547926/work
packaging @ file:///tmp/build/80754af9/packaging_1607971725249/work
pluggy==0.13.1
py @ file:///tmp/build/80754af9/py_1607971587848/work
pycryptodome==3.3.1
pylint==2.3.1
pyparsing==2.4.7
pytest==0.0.0
python-jose-cryptodome==1.3.2
six @ file:///C:/ci/six_1605205426665/work
SQLAlchemy==1.3.3
tfds-nightly==3.1.0.dev202006160105
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
typed-ast==1.3.5
Werkzeug==0.15.2
wincertstore==0.2
wrapt==1.11.1
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
